+++ mllib-dal/src/main/native/OneCCL.h
@@ -44,4 +44,4 @@ event CCL_API gather(const BufferType *sendbuf, int sendcount,
 
 ccl::communicator &getComm();
 ccl::shared_ptr_class<ccl::kvs> &getKvs();
-extern const int ccl_root;
+extern const size_t ccl_root;
